chore: extracted out common methods to a mailbox helper
was forced to do this because code-climate raised this
as identical blocks of code in two files and asked to
refactor
